Ensure selectors are stringified
In code like

    my $page_four = $drv->find(4, method => 'link_text');

the value 4 was previously passed to the JSON serialiser as-is, which
may well choose to encode it as a JSON number rather than as a JSON
string. This appears to fail in ChromeDriver. Fix this by stringifying
whatever the user wants to find.

I can't see a good way to write a test for this behaviour,
unfortunately.