Better reporting for certain sorts of error
For some erroneous uses of WebDriver, I've seen ChromeDriver responses
that look like this:

  {
    sessionId => "ee2b09c20c3309656e5980ea33fba872",
    status => 10,
    value => {
      message => "stale element reference: element is not attached to the page document\n  (Session info: chrome=54.0.2840.100)\n  (Driver info: chromedriver=2.25.426924 (649f9b868f6783e496e78c123212b908bf3b232e),platform=Linux 4.4.0-47-generic x86_64)"
    }
  }

Since $reply->{value} is a hash rather than an array in this case, the
attempt to iterate over it as an array failed with an obscure error
message.

With this change, there's still a failure and an error message, but
hopefully it is less obscure.